update dockerfile to handle *.gemspec on build
Because of the addition of the *.gemspec file into the build
process, it needs to be included in the Dockerfile to correctly
install all the gems.

We also don't need to include the `spec.files` lines, so removing
those.

And lastly, since we're not entering into the running container,
also removing the `-it` from the docker run command in the README.